src/requirements.txt
Flask
Flask-SocketIO
requests
python-dotenv 